refactor: RepositorySection → RepositoryColumn 컬럼 컴포넌트 분리
RepositorySection에 인라인되어 있던 레포지토리 컬럼 렌더링 로직을
독립 컴포넌트 RepositoryColumn으로 추출

- RepositoryColumn.tsx: 단일 컬럼 레이아웃 및 아코디언 렌더링 담당
- RepositorySection.tsx: 데이터 페칭 및 컬럼 조합 역할로 축소
